backend/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.5
flask-cors==4.0.0
flask-session==0.5.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
openai==1.3.0
gunicorn==21.2.0